LayoutTests/fast/canvas/webgl/buffer-data-array-buffer-expected.txt

Test bufferData/bufferSubData with ArrayBuffer input

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=41884 : Implement bufferData and bufferSubData with ArrayBuffer as input
PASS gl is non-null.
PASS array is non-null.
PASS [object WebGLBuffer] is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
